OnPoint-014 Move crashing junit test to instrumented tests [GM]
The crash happens before the test is even run, when Junit tries to
load the proper Android framework classes but does not find them.
It's probably an issue in the run configuration, but there is no
time to debug it
